requirements.txt
fastapi
uvicorn[standard]
httpx
pydantic
pydantic-settings
google-genai
python-dotenv
requests